+++ LayoutTests/fast/element-targeting/target-reinserted-elements-expected.html
@@ -1,0 +1,19 @@
+<!DOCTYPE html> <!-- webkit-test-runner [ useFlexibleViewport=true ] -->
+<html>
+<meta name="viewport" content="width=device-width, initial-scale=1">
+<head>
+<style>
+body, html {
+    margin: 0;
+}
+
+.content {
+    width: 320px;
+    height: 200px;
+}
+</style>
+</head>
+<body>
+<p class="content">This test requires WebKitTestRunner.</p>
+</body>
+</html>
